Rename Coord to RectangleExtension & Change purpose model class to util class
Coord.java
	delete constructor
	delete inheritance Rectangle
	change scale method to static method

Before changing class, this class is useless except scale function
So, change Coord.java to RectangleExtension.class for using util class
